% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/river_networks.R
\name{add.node}
\alias{add.node}
\title{Add a new node in between two nodes in an igraph graph}
\usage{
add.node(g, from, to)
}
\arguments{
\item{g}{an igraph graph}

\item{from}{a vector of node IDs}

\item{to}{a vector of node IDs}
}
\value{
an igraph graph
}
\description{
This increases the size of the graph by the number of new nodes
}
\details{
Description - This function does stuff.
}

