% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/river_networks.R
\name{buildTopo}
\alias{buildTopo}
\title{Create an igraph graph from a spatial lines data.frame network}
\usage{
buildTopo(lines, from = "FNODE_", to = "TNODE_")
}
\arguments{
\item{lines}{a spatial lines data.frame containing a river network}

\item{from}{the column name containing the 'from' node IDs}

\item{to}{the column name containing the 'to' node IDs}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}

