% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avg_diff.R
\name{calc_avg_diff}
\alias{calc_avg_diff}
\title{Calculate (row wise) mean difference from best}
\usage{
calc_avg_diff(x)
}
\arguments{
\item{d}{}
}
\value{
A column with the avg difference from the best
}
\description{
Calculate (row wise) mean difference from best
}
\examples{
calc_avg_diff(d)
}
