% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_best.R
\name{calc_best}
\alias{calc_best}
\title{Calculate best rate}
\usage{
calc_best(x)
}
\arguments{
\item{d}{}
}
\value{
A column with the best rate
}
\description{
Calculate best rate
}
\examples{
calc_best(d)
}
