% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate difference from the best}
\usage{
calc_diff(x)
}
\arguments{
\item{d}{}
}
\value{
Raced columns with difference from the best
}
\description{
Calculate difference from the best
}
\examples{
calc_diff(d)
}
