% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_id.R
\name{calc_id}
\alias{calc_id}
\title{Calculate index of disparity}
\usage{
calc_id(x)
}
\arguments{
\item{d}{}
}
\value{
An index of disparity column
}
\description{
Calculate index of disparity
}
\examples{
calc_id(d)
}
