% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_var.R
\name{calc_p_var}
\alias{calc_p_var}
\title{Calculate (row wise) POPULATION variance of differences from best - use for non-sample data like Decennial Census, CDPH Births, or CADOJ Incarceration}
\usage{
calc_p_var(x)
}
\arguments{
\item{d}{}
}
\value{
A variance column
}
\description{
Calculate (row wise) POPULATION variance of differences from best - use for non-sample data like Decennial Census, CDPH Births, or CADOJ Incarceration
}
\examples{
calc_p_var(d)
}
