% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ranks.R
\name{calc_ranks}
\alias{calc_ranks}
\title{Calculate disparity and performance ranks and quadrants}
\usage{
calc_ranks(x)
}
\arguments{
\item{d}{}
}
\value{
A column with disparity rank, performance rank, and a quadrant color
}
\description{
Calculate disparity and performance ranks and quadrants
}
\examples{
calc_ranks(d)
}
