% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_s_var.R
\name{calc_s_var}
\alias{calc_s_var}
\title{Calculate (row wise) SAMPLE variance of differences from best - use for sample data like ACS or CHIS}
\usage{
calc_s_var(x)
}
\arguments{
\item{d}{}
}
\value{
A variance column
}
\description{
Calculate (row wise) SAMPLE variance of differences from best - use for sample data like ACS or CHIS
}
\examples{
calc_state_z(state_table)
}
