% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_state_z.R
\name{calc_state_z}
\alias{calc_state_z}
\title{Calculate State Z-scores}
\usage{
calc_state_z(x)
}
\arguments{
\item{state_table}{}
}
\value{
Raced disparity z score columns
}
\description{
Calculate State Z-scores
}
\examples{
calc_state_z(state_table)
}
