% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_z.R
\name{calc_z}
\alias{calc_z}
\title{Calculate County Disparity Z-scores}
\usage{
calc_z(x)
}
\arguments{
\item{county_table}{}
}
\value{
Raced disparity z score columns
}
\description{
Calculate County Disparity Z-scores
}
\examples{
calc_z(county_table)
}
