% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_to_postgres.R
\name{city_to_postgres}
\alias{city_to_postgres}
\title{Send city table to Postgres}
\usage{
city_to_postgres(x)
}
\arguments{
\item{city_table}{}
}
\value{
Postgres table
}
\description{
Send city table to Postgres
}
\examples{
to_postgres(county_table) or to_postgres(state_table)
}
