% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_values.R
\name{count_values}
\alias{count_values}
\title{Count the number of rate values}
\usage{
count_values(x)
}
\arguments{
\item{D}{}
}
\value{
A column with values_count
}
\description{
Count the number of rate values
}
\examples{
d <- count_values(d)
}
