% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rc_Qfuncs.R
\name{Dgraph}
\alias{Dgraph}
\title{Differences defining a river network}
\usage{
Dgraph(g)
}
\arguments{
\item{g}{an igraph graph specifying the dependence structure}
}
\value{
a Matrix with a column for each node and a row for each connection / edge
}
\description{
Returns the incidence matrix of a graph
}
\details{
Description - This function does stuff.
}

