% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir.R
\name{eapply_fast}
\alias{eapply_fast}
\title{apply a function f to all elements in a hashed envir for side-effects (no return value)}
\usage{
eapply_fast(env, fn, ...)
}
\description{
apply a function f to all elements in a hashed envir for side-effects (no return value)
}
\examples{
\dontrun{
e <- new.env()
for(a in LETTERS[1:10]){
  assign(x = a,value = rpois(n = 5,lambda = 10),envir = e)
 }
 fn <- function(x,msg){
 cat(msg,"\\n")
 cat(x,"\\n")
 }
eapply_fast(env = e,fn = fn,msg = "hi")
}
}
