% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir.R
\name{get_fromEnv}
\alias{get_fromEnv}
\title{return a named element of an environment}
\usage{
get_fromEnv(name, env)
}
\description{
this function's source is actually in utilities.h/c
}
\examples{
\dontrun{
e <- new.env(hash = T)
e$thing <- rpois(n = 5,lambda = 10)
get_fromEnv("thing",env)
}
}
