% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rc_Qfuncs.R
\name{getQgraph}
\alias{getQgraph}
\title{Compute RW1 precision matrix on a river network}
\usage{
getQgraph(g, weights = NULL)
}
\arguments{
\item{g}{an igraph graph specifying the dependence structure}

\item{weights}{weights to be applied to the edges of the graph (see details)}
}
\value{
what does it return
}
\description{
Returns the laplacian matrix of a graph.  If weights are supplied
then the weighted lacplacian is returned.  Sensible weights are
based on the flows of the incoming upstream graph edges.
}
\details{
Weights argument is passed onto the graph.laplacian function from
the igraph package. If weights is NULL (default) and the graph has
an edge attribute called weight, then it will be used automatically.
Set this to NA if you want the unweighted Laplacian on a graph that
has a weight edge attribute.
}

