% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCR.R
\name{mcmcMH}
\alias{mcmcMH}
\title{MCR model Random Walk Metropolis-Hastings MCMC (interface to C)}
\usage{
mcmcMH(initTheta, proposalSD, numIterations, GFPp_ym_F_1, GFPp_yp_F_1,
  GFPm_ym_F_1, GFPm_yp_F_1, GFPp_ym_M_1, GFPp_yp_M_1, GFPm_ym_M_1,
  GFPm_yp_M_1, GFPp_ym_F_2, GFPp_yp_F_2, GFPm_ym_F_2, GFPm_yp_F_2,
  GFPp_ym_M_2, GFPp_yp_M_2, GFPm_ym_M_2, GFPm_yp_M_2, GFPp_ym_F_3,
  GFPp_yp_F_3, GFPm_ym_F_3, GFPm_yp_F_3, GFPp_ym_M_3, GFPp_yp_M_3,
  GFPm_ym_M_3, GFPm_yp_M_3, GFPp_ym_F_4, GFPp_yp_F_4, GFPm_ym_F_4,
  GFPm_yp_F_4, GFPp_ym_M_4, GFPp_yp_M_4, GFPm_ym_M_4, GFPm_yp_M_4, gens)
}
\description{
MCR model Random Walk Metropolis-Hastings MCMC (interface to C)
}
