\name{RC.capture}
\Rdversion{1.1}
\alias{RC.capture}
\title{
Capture expression
}
\description{
Captures an R expression and returns RC.resultText. For internal use only.
}
\usage{
RC.capture(expression, collapse = NULL)
}
\arguments{
  \item{expression}{
an R expression
}
  \item{collapse}{
optional character to separate the results
}
}
\keyword{ internal }
