\name{RC.feedback.submitted}
\Rdversion{1.1}
\alias{RC.feedback.submitted}
\title{
Number of Submitted Feedback Messages
}
\description{
This function computes the # of feedback messages that have been submitted by a user in the forums of FreeStatistics.org.
}
\usage{
RC.feedback.submitted(id = "", keyword = "", quote = TRUE)
}
\arguments{
  \item{id}{
user id
}
  \item{keyword}{
optional keyword to be used in the search of computations
}
  \item{quote}{
boolean parameter which indicates if quotes are used in the search
}
}
\details{
This function is not yet available.
}
%%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%%}
%%\examples{
%%
%%}
\keyword{ misc }

