\name{RC.hello}
\Rdversion{1.1}
\alias{RC.hello}
\title{
Say Hello to the RC network
}
\description{
This function tests the network connectivity and performance of the R servers that are connected to the R Framework. This function is mainly useful whenever the user wishes to determine is errors or problems are related to the internet connection or the availability of R servers.
}
\usage{
RC.hello()
}
\examples{
#RC.hello()
}
\keyword{ utilities }

