\name{RC.igraph}
\alias{RC.igraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert Network into Igraph object
}
\description{
This function converts a network object that is generated by the RC.network() function and converts it into an igraph object (from the igraph package) with additional attributes.
}
\usage{
RC.igraph(network = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{network}{
network object as generated by RC.network() or RC.network.list()
}
}
\value{
The function returns an igraph object (see igraph package for more information).
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\seealso{
\code{\link{igraph}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }

