\name{RC.igraph.union}
\alias{RC.igraph.union}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create the Union of two Sociograms
}
\description{
This function joins two network objects and returns the corresponding union igraph object (from the igraph package).
}
\usage{
RC.igraph.union(r1, r2, dovert = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{r1}{
the first network object as generated by RC.network() or RC.network.list()
}
  \item{r2}{
the second network object as generated by RC.network() or RC.network.list()
}
  \item{dovert}{
boolean
}
}
\value{
The function returns an igraph object (see igraph package for more information).
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }

