\name{RC.print.html}
\Rdversion{1.1}
\alias{RC.print.html}
\title{
Print the HTML output of a computation
}
\description{
This function converts the 'a' object into a readable array for the R console. The 'a' object is created by the RC.reproduce() function whenever a computation is involved that was originally generated by an R module (published on the web).
}
\usage{
RC.print.html(a)
}
\arguments{
  \item{a}{
This variable contains the raw output of the computation (only applies to computations that generate result tables in HTML format)
}
}
\details{
This function converts the tables of the HTML output into R objects (arrays) and prints them in the console.
}
%%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%%}
\examples{
#r <- RC.ls(keyword="airline")
#RC.reproduce(r$url[40])
#convert the HTML output and print it
#RC.print.html(a)
}
\keyword{ print }

