% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funprog.R
\name{Reduce_Simple}
\alias{Reduce_Simple}
\title{Simple version of "Reduce"}
\usage{
Reduce_Simple(f, x)
}
\description{
simple version of Reduce (see http://adv-r.had.co.nz/Functionals.html#functionals-fp)
}
\examples{
\dontrun{

}
}
