% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{sub_array3}
\alias{sub_array3}
\title{subset a 3d array}
\usage{
sub_array3(i, j, k, array)
}
\description{
subset a 3d array
}
\examples{
\dontrun{
 x <- array(data = 1:(2*3*4),dim = c(2,3,4))
x[1,3,4]
sub_array3(1,3,4,x)
}

}
