% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{sub_array4}
\alias{sub_array4}
\title{subset a 4d array}
\usage{
sub_array4(i, j, k, l, array)
}
\description{
subset a 4d array
}
\examples{
\dontrun{
 x <- array(data = 1:(2*3*4*5),dim = c(2,3,4,5))
x[1,3,4,5]
sub_array4(1,3,4,5,x)
}

}
