% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copymat.R
\name{submat}
\alias{submat}
\title{Subset Matrix in C by rows}
\usage{
submat(mat, nrow)
}
\description{
Subset Matrix in C by rows
}
\examples{
\dontrun{
   mat <- matrix(1:15,nrow=5,ncol=3)
   submat(mat,2)
}
}
