% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copymat.R
\name{submat_cols}
\alias{submat_cols}
\title{Subset Matrix in C by cols}
\usage{
submat_cols(mat, ncol)
}
\description{
Subset Matrix in C by cols
}
\examples{
\dontrun{
   mat <- matrix(1:15,nrow=3,ncol=5)
   submat_cols(mat,2)
}
}
