% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_postgres.R
\name{to_postgres}
\alias{to_postgres}
\title{Send county and state tables to Postgres}
\usage{
to_postgres(x, y)
}
\arguments{
\item{county_table}{or state_table}
}
\value{
Postgres table
}
\description{
Send county and state tables to Postgres
}
\examples{
to_postgres(county_table) or to_postgres(state_table)
}
