% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_values.R
\name{count_values}
\alias{count_values}
\title{Calculate Number of Raced Rate Values}
\usage{
count_values(x)
}
\arguments{
\item{d}{}
}
\value{
A column with a rate value
}
\description{
Calculate Number of Raced Rate Values
}
\examples{
count_values(d)
}
