% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC-package.R
\docType{package}
\name{RC-package}
\alias{RC}
\alias{RC-package}
\title{RC: Standard functions for RaceCounts analysis}
\description{
The package contains the most common user-defined functions that the Research & Data Team use for data analysis in RaceCounts. For example, the functions calculates the county and state disparity z-scores, disparity and performance ranks, and quadrant colors.
}
\author{
\strong{Maintainer}: Catalyst California \email{cringewald@catalystcalifornia.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
