\name{RC.powered.by}
\Rdversion{1.1}
\alias{RC.powered.by}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Powered by?
}
\description{
This function advertises Fedora linux as a free and robust OS. All servers of the FreeStatistics.org and Wessa.net network run Fedora linux.
}
\usage{
RC.powered.by()
}
\references{
http://www.fedoraproject.org/
}
\keyword{ print }

